+++ .pre-commit-config.yaml
@@ -24,7 +24,7 @@
     - id: trailing-whitespace
     - id: detect-private-key
     - id: flake8
-      args: ['--ignore=W503', '--max-line-length=150']
+      args: ['--ignore=W503', '--max-line-length=250']
     - id: trailing-whitespace
       args: [--markdown-linebreak-ext=md]
 - repo: https://github.com/ambv/black
