+++ shared_code/pact/opg_pact/docker-compose.yml
@@ -55,4 +55,3 @@ services:
         default:
           aliases:
             - local.mock
-
